refactor(md): R0-R4 — critical fixes, dedup erfc, f64 migration, robustness
R0 Critical fixes:
- Fix linked list corruption in CellList3D (duplicate insertion lines)
- Remove LJ force clipping anti-pattern (.clamp hides overflow)
- Fix spatial_tree particle loss in degenerate cells (count increment)

R1 Consolidate constants:
- Create special_functions.rs: shared erfc_approx (was duplicated in
  ewald.rs + thermostat.rs), named A&S 7.1.26 coefficients
- Named constants: NUMERICAL_GRAD_STEP, DEG_TO_RAD, LJ_SOFTENING_RADIUS_SQ
- Replace hardcoded TIP3P geometry in constraints.rs with water.rs imports

R2 Performance (partial):
- Cache native_mask in GoTopology (eliminates N² alloc per force call)
- Remove build_native_mask() from remd.rs and parallel_forces.rs
- Analytical gradients deferred (ADR-028, optimization not correctness)

R3 f64 migration:
- BondParams/AngleParams/DihedralParams: f32 → f64
- bonded.rs: all functions now f64 (eliminate f32↔f64 casts)
- Cascade through amber.rs, water.rs, constraints.rs, bonded_forces.rs,
  peptide_vacuum.rs, peptide_solvated.rs

R4 Robustness:
- SHAKE non-convergence diagnostic (debug_assertions)
- Guard debug_assert on unknown amino acid types in go_model

Tests: 3593 passing (0 regressions, +5 from migration validation).

+++ src/batch/ff/amber.rs
@@ -119,10 +119,10 @@ fn parse_bond_line(
 ) -> Result<(), String> {
     let (types_str, rest) = split_types_and_values(line, 2, line_num, "BOND")?;
     let types = parse_type_pair(&types_str)?;
-    let k: f32 = rest[0]
+    let k: f64 = rest[0]
         .parse()
         .map_err(|_| format!("line {}: bad bond k: {}", line_num + 1, rest[0]))?;
-    let r0: f32 = rest[1]
+    let r0: f64 = rest[1]
         .parse()
         .map_err(|_| format!("line {}: bad bond r_eq: {}", line_num + 1, rest[1]))?;
     ff.bond_params.push((types.0, types.1, BondParams { r0, k }));
@@ -135,13 +135,13 @@ fn parse_angle_line(
 ) -> Result<(), String> {
     let (types_str, rest) = split_types_and_values(line, 2, line_num, "ANGLE")?;
     let types = parse_type_triplet(&types_str)?;
-    let k: f32 = rest[0]
+    let k: f64 = rest[0]
         .parse()
         .map_err(|_| format!("line {}: bad angle k: {}", line_num + 1, rest[0]))?;
     let theta_deg: f64 = rest[1]
         .parse()
         .map_err(|_| format!("line {}: bad angle theta: {}", line_num + 1, rest[1]))?;
-    let theta0 = (theta_deg * std::f64::consts::PI / 180.0) as f32;
+    let theta0 = theta_deg * std::f64::consts::PI / 180.0;
     ff.angle_params.push((types.0, types.1, types.2, AngleParams { theta0, k }));
     Ok(())
 }
@@ -165,8 +165,8 @@ fn parse_dihe_line(
         .parse()
         .map_err(|_| format!("line {}: bad dihedral periodicity: {}", line_num + 1, rest[3]))?;
 
-    let k = if divider.abs() > 1e-12 { (barrier / divider) as f32 } else { barrier as f32 };
-    let delta = (phase_deg * std::f64::consts::PI / 180.0) as f32;
+    let k = if divider.abs() > 1e-12 { barrier / divider } else { barrier };
+    let delta = phase_deg * std::f64::consts::PI / 180.0;
     let n = periodicity.abs() as u8;
 
     ff.dihedral_params.push((types.0, types.1, types.2, types.3, DihedralParams { k, n, delta }));
@@ -325,7 +325,7 @@ O     1.661    0.2100
         assert!((ct_ct_n.k - 80.0).abs() < 1e-6);
         // 109.7 degrees in radians
         let expected = 109.7_f64 * std::f64::consts::PI / 180.0;
-        assert!((ct_ct_n.theta0 as f64 - expected).abs() < 1e-3);
+        assert!((ct_ct_n.theta0 - expected).abs() < 1e-3);
     }
 
     #[test]
